[3.9] Check whether process is used in ImportConfiguration before deletion (#7010)
* [3.9] Check whether process is used in ImportConfiguration before deletion

* Implement change request
