+++ Kitodo/src/main/java/org/kitodo/production/forms/ProcessForm.java
@@ -92,7 +92,6 @@ public class ProcessForm extends TemplateBaseForm {
 
     private String processEditReferer = DEFAULT_LINK;
     private String taskEditReferer = DEFAULT_LINK;
-    private String errorMessage = "";
 
     private List<SelectItem> customColumns;
 
@@ -1198,22 +1197,6 @@ public class ProcessForm extends TemplateBaseForm {
     }
 
     /**
-     * Rename media files of all selected processes.
-     */
-    public void renameMedia() {
-        Stopwatch stopwatch = new Stopwatch(this, "renameMedia");
-        List<Process> processes = getSelectedProcesses();
-        errorMessage = ServiceManager.getFileService().tooManyProcessesSelectedForMediaRenaming(processes.size());
-        if (StringUtils.isBlank(errorMessage)) {
-            PrimeFaces.current().executeScript("PF('renameMediaConfirmDialog').show();");
-        } else {
-            Ajax.update("errorDialog");
-            PrimeFaces.current().executeScript("PF('errorDialog').show();");
-        }
-        stopwatch.stop();
-    }
-
-    /**
      * Start renaming media files of selected processes.
      */
     public void startRenaming() {
@@ -1234,15 +1217,6 @@ public class ProcessForm extends TemplateBaseForm {
         Stopwatch stopwatch = new Stopwatch(this, "getMediaRenamingConfirmMessage");
         return stopwatch.stop(Helper.getTranslation("renameMediaForProcessesConfirmMessage", String.valueOf(
             getSelectedProcesses().size())));
-    }
-
-    /**
-     * Get error message.
-     * @return error message
-     */
-    public String getErrorMessage() {
-        Stopwatch stopwatch = new Stopwatch(this, "getErrorMessage");
-        return stopwatch.stop(errorMessage);
     }
 
     /**
