+++ Kitodo/src/main/resources/messages/errors_de.properties
@@ -180,6 +180,7 @@ processTitleAlreadyInUse=Der Prozesstitel ''{0}'' wird bereits verwendet.
 processTitleEmpty=Kein Vorgangstitel angegeben.
 processTitleInvalid=Ungültiger Titel für einen Vorgang. Er sollte mit dem regulären Ausdruck ''{0}'' übereinstimmen.
 projectTitleAlreadyInUse=Der Projekttitel wird bereits verwendet.
+processUsedAsTemplateProcess=Vorgang kann nicht gelöscht werden, da er in mindestens einer Importkonfiguration als Vorlagenvorgang verwendet wird: ''{0}''
 propertiesForProcessNotSaved=Eigenschaften des Prozesses {0} konnten nicht gespeichert werden.
 
 # R
