+++ Kitodo/src/main/resources/messages/errors_en.properties
@@ -180,6 +180,7 @@ processTitleAlreadyInUse=The process title ''{0}'' is already in use.
 processTitleEmpty=No process title stated.
 processTitleInvalid=Invalid title for process. It should match ''{0}'' regular expression.
 projectTitleAlreadyInUse=The project title is already in use.
+processUsedAsTemplateProcess=Process cannot be deleted because it is used as a template process in at least one Import Configuration: ''{0}''
 propertiesForProcessNotSaved=Properties for process {0} could not be saved.
 
 # R
