+++ Kitodo/src/main/resources/messages/errors_es.properties
@@ -188,6 +188,8 @@ processTitleAlreadyInUse=El título del proceso ''{0}'' ya está en uso.
 processTitleEmpty=No se especifica el título del proceso.
 processTitleInvalid=Título del proceso inválido. Debe coincidir con la expresión regular ''{0}''.
 projectTitleAlreadyInUse=El título del proyecto ya está en uso.
+# translated by Google, please check!
+processUsedAsTemplateProcess=El proceso no se puede eliminar porque se utiliza como proceso plantilla en al menos una configuración de importación: ''{0}''
 propertiesForProcessNotSaved=Fallo al guardar las propiedades del proceso {0}.
 
 # R
